pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "n2bl-batch-yt-uploader"
version = "0.1.0"
description = "Batch YouTube uploader with throttling and sidecar metadata"
requires-python = ">=3.8"
dependencies = [
  "google-api-python-client",
  "google-auth",
  "google-auth-oauthlib",
  "google-auth-httplib2",
  "tabulate",
  "colorama",
  "tqdm",
]

[project.scripts]
pqp = "n2bl.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
